Move AppPreviewCard to platform, and leave a wordless scheduled message alone
Two follow-ups.

`apps/ai_models/playground/PlaygroundApps.tsx` imports AppPreviewCard, which
lived in apps/builder — an app importing another app, which
check-boundaries.mjs refuses. That check runs first in `npm run build`, so
nothing built and dev.sh could not boot. The card imports nothing but
@platform and is drawn by /apps, /home and now the playground, so it belongs
in platform/ui. Pure move: the file's contents are unchanged and the three
importers plus two tests that read it by path are repointed.

And `_outgoing` now leaves a message with no words exactly as it is. The
app-state block is a prefix to something a human said; on its own it is a
send that is pure machinery, which every reader strips back to "" — a turn
whose only content is a description of a file.
