+++ frontend/src/platform/ui/AppPreviewCard.tsx
